Safari-609.1.11.1/LayoutTests/accessibility/ios-simulator/slider-min-value-crash.html

<!DOCTYPE HTML PUBLIC "-//IETF//DTD HTML//EN">
<html>
<head>
<script src="../../resources/js-test-pre.js"></script>
</head>
<body id="body">

<input type="range" id="slider" value="50" min="25" max="75">

<p id="description"></p>
<div id="console"></div>

<script>

    description("This tests that accessing the min/max value of a detached ax object won't cause crash.");

    if (window.accessibilityController) {

        var sliderObj = accessibilityController.accessibleElementById("slider");
        shouldBe("sliderObj.minValue", "25");
        
        // Remove the element
        var slider = document.getElementById("slider");
        slider.parentNode.removeChild(slider);
        
        // Make sure it won't crash
        shouldBe("sliderObj.minValue", "0");
    }

</script>

<script src="../../resources/js-test-post.js"></script>
</body>
</html>
